Refactor test code to use new assertion
To simplify the last two test cases, create a custom assertion for
determining if two Locations are not equal.

NOTE: This is implemented as an extension to the XCTestCase class
rather than adding to LocationTests class. This helps with reading the
test case class by having all the test cases within the class, and
other utility functions as extensions.

This could be further enhanced by creating this extension in a
seperate file containing other custom assertions.