Deque table view cells
Use a mock object to enable verirfication the method
dequeueReusableCell(_:_) is called. This is used to drive
implementation of cell(forRowAt indexPath), where the cell is dequed.

NOTE: This causes the test: testCellForRowReturnsItemCell() to fail
with an exception as the new implementation of cell(forRowAt indexPath)
requires a cell identifier to deque cells, and this has not be
registered.