Add table view cell with identifier
To ensure the table view tableView(cellForRowAt indexPath:) method
deques cells with a known identifier, add a UITableViewCell to the
storyboard with the identifier 'ItemCell'.

The test code is refactored to get access to the storyboard, and from
this the table view which contains this 'ItemCell'. This ensures the
test testCellForRowReturnsItemCell() now passes.

NOTE: The test testCellForRowDequeuesCellFromTableView() must still
register a cell as it is using the MockTableView rather than the table
view from the storyboard.