Add configCell() method to ItemCell
To configure UI items in an ItemCell, data needs to be passed to the
cell during tableView(cellForRowAt:). To achieve this, a configCell()
method is created which will be used to update UI items (in this case
text labels) for the ItemCell.

Since there is no other way to determine if the method has been called,
a mock ItemCell is created to test this.