Add check button to DetailViewController
The approach used here is different to the book. Instead, we use the
approach described by Jon Reid in his screencast "How to Do
UIViewController TDD". In Jon's approach, he explains when dealing with
a UIButton there are three things that need to be verified:
 1. Is the outlet hooked up?
 2. Is the outlet connected to an action?
 3. Does the action do what it is supposed to do?

The first of these may not be necessary, as in most iOS development an
outlet for a button is not created - only an action. However, this
explicitly tests for the existence of a button.

In the book "Test-Driven iOS Development with Swift 4", only the method
checkItem() is tested to verify it updates the ToDo and Done counts.
Indeed, the button iteself is not actually added to the
DetailViewController - let alone tested.

We should also test the button label is correct. This will be done
next.

Some changes to constraints have also been made.