Refactor tests for existence of labels
Rather than use the approach described in the book "Test-Driven iOS
Development with Swift 4" of testing if a label is a descendant of the
view controller view, simply test that it is a property of the view
controller. This achieves the same verification, however, it uses a
simpler and easier to understand approach.