Add item manager and action for save button
This is somewhat different to the text "Test-Driven iOS Development
with Swift 4" in that we test the saveItem() method saves a new
ToDoItem to a ItemManager which is the model for this view cotroller.

These test only cover the title, timestamp and description fields. The
location field will be added using a mock geocoder - which is described
in "Test-Driven iOS Development with Swift 4".